feat: Enhance checkout process with coupon functionality and discount calculations
- Added coupon model and integrated coupon validation in CheckoutController.
- Implemented discount calculations based on coupon type (percentage or fixed) and minimum amount requirements.
- Updated product request validation to include new boolean fields for product status.
- Modified Coupon model to cast coupon_type and discount fields to appropriate types.
- Enhanced Product and ProductVariant models with decimal casting for price-related fields.
- Updated Order and OrderItems repositories to handle coupon application and discount amounts.
- Improved product creation and update processes to include new status fields.
- Added visibility flags for products in the admin panel.
- Enhanced frontend views for checkout and order summaries to display subtotal, discounts, and grand totals.
- Updated migrations to reflect new fields and types in the database schema.
- Improved customer and product management views to display active status.
- Refactored JavaScript for cart and checkout to dynamically calculate totals based on discounts.

+++ database/migrations/2026_06_20_144232_create_coupons_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('coupon_code');
             $table->string('coupon_type')->nullable();
-            $table->float('discount')->nullable()->default(0);
-            $table->float('min_amount')->nullable()->default(0);
+            $table->decimal('discount', 10, 2)->nullable()->default(0);
+            $table->decimal('min_amount', 10, 2)->nullable()->default(0);
             $table->integer('limit')->nullable()->default(0);
             $table->integer('total_applied')->nullable()->default(0);
             $table->timestamp('start_date')->nullable();
